Stripe Payment Gateway added

+++ src/main/java/com/project/Airbnb/service/impl/HotelServiceImpl.java
@@ -5,7 +5,9 @@ import com.project.Airbnb.dto.HotelInfoDto;
 import com.project.Airbnb.dto.RoomDto;
 import com.project.Airbnb.entity.Hotel;
 import com.project.Airbnb.entity.Room;
+import com.project.Airbnb.entity.User;
 import com.project.Airbnb.exception.ResourceNotFoundException;
+import com.project.Airbnb.exception.UnAuthorizedException;
 import com.project.Airbnb.repository.HotelRepository;
 import com.project.Airbnb.repository.RoomRepository;
 import com.project.Airbnb.service.HotelService;
@@ -14,6 +16,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,6 +35,9 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating new hotel with name {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);//by default
+
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -43,6 +49,10 @@ public class HotelServiceImpl implements HotelService {
         log.info("Getting the hotel with ID :{}", id);
         Hotel hotel = hotelRepository.findById(id).
                 orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: {}" + id));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("this user does not own this hotel with id "+ id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -52,10 +62,15 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("No hotel found with id " + id)
         );
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("this user does not own this hotel with id "+ id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
         log.info("Hotel updated with id: {}", hotel.getId());
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -65,6 +80,10 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("No hotel found with id " + id)
         );
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("this user does not own this hotel with id "+ id);
+        }
         //TODO deletet the future inventory for this hotel
         for (Room room : hotel.getRooms()) {
             inventoryService.deleteFutureInventories(room);
@@ -79,8 +98,14 @@ public class HotelServiceImpl implements HotelService {
     @Override
     public void activateHotel(Long id) {
         log.info("Activating the hotel with ID :{}", id);
+
         Hotel hotel = hotelRepository.findById(id).
                 orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: {}" + id));
+
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("this user does not own this hotel with id "+ id);
+        }
         hotel.setActive(true);
         //TODO create inventory for all the room for this hotels
