+++ src/main/java/com/project/Airbnb/service/impl/RoomServiceImpl.java
@@ -3,7 +3,9 @@ package com.project.Airbnb.service.impl;
 import com.project.Airbnb.dto.RoomDto;
 import com.project.Airbnb.entity.Hotel;
 import com.project.Airbnb.entity.Room;
+import com.project.Airbnb.entity.User;
 import com.project.Airbnb.exception.ResourceNotFoundException;
+import com.project.Airbnb.exception.UnAuthorizedException;
 import com.project.Airbnb.repository.HotelRepository;
 import com.project.Airbnb.repository.RoomRepository;
 import com.project.Airbnb.service.InventoryService;
@@ -11,6 +13,7 @@ import com.project.Airbnb.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +34,10 @@ public class RoomServiceImpl implements RoomService {
         log.info("Creating a new room in hotel with ID: {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() ->
                 new ResourceNotFoundException("No hotel found with id " + hotelId));
+//        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+//        if (!user.equals(hotel.getOwner())){
+//            throw new UnAuthorizedException("this user does not own this hotel with id "+ hotelId);
+//        }
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -65,6 +72,11 @@ public class RoomServiceImpl implements RoomService {
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorizedException("this user does not own this hotel with id "+ roomId);
+        }
         inventoryService.deleteFutureInventories(room);
         roomRepository.deleteById(roomId);
     }
